pijuice/requirements.txt
balena-sdk
paho-mqtt
python-dateutil
twilio
w1thermsensor
